# refactor: improve the performance of block processing