--- a/benches/benches/process_block.rs
+++ b/benches/benches/process_block.rs
@@ -1,5 +1,5 @@
 use ckb_chain::chain::{ChainBuilder, ChainController};
-use ckb_chain_spec::consensus::Consensus;
+use ckb_chain_spec::consensus::{Consensus, ProposalWindow};
 use ckb_core::block::{Block, BlockBuilder};
 use ckb_core::header::HeaderBuilder;
 use ckb_core::script::Script;
@@ -20,139 +20,162 @@ use std::sync::Arc;
 use tempfile::{tempdir, TempDir};
 
 fn bench(c: &mut Criterion) {
-    // benchmark processing 5 blocks on main branch
-    // 0 -> 1 -> 2 -> 3 -> 4 -> 5
-    c.bench_function("main_branch", |b| {
-        b.iter_with_setup(
-            || {
-                let (chain, shared, dir) = new_chain();
-                let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
-                (0..5).for_each(|_| {
-                    let parent_index = blocks.len() - 1;
-                    gen_block(&mut blocks, parent_index);
-                });
-                (chain, blocks, dir)
-            },
-            |(chain, blocks, _dir)| {
-                blocks.into_iter().skip(1).for_each(|block| {
-                    chain
-                        .process_block(Arc::new(block))
-                        .expect("process block OK")
-                });
-            },
-        )
-    });
+    let txs_sizes = vec![100usize, 200, 500, 1000];
+
+    // benchmark processing 20 blocks on main branch
+    c.bench_function_over_inputs(
+        "main_branch",
+        |b, txs_size| {
+            b.iter_with_setup(
+                || {
+                    let (chain, shared, dir, system_cell_hash, data_hash) = new_chain(*txs_size);
+                    let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
+                    (0..20).for_each(|_| {
+                        let parent_index = blocks.len() - 1;
+                        gen_block(&mut blocks, parent_index, &system_cell_hash, &data_hash);
+                    });
+                    (chain, blocks, dir)
+                },
+                |(chain, blocks, _dir)| {
+                    blocks.into_iter().skip(1).for_each(|block| {
+                        chain
+                            .process_block(Arc::new(block))
+                            .expect("process block OK")
+                    });
+                },
+            )
+        },
+        txs_sizes.clone(),
+    );
 
     // benchmark processing 2 blocks on side branch
     // 0 -> 1 -> 2 -> 3 -> 4 -> 5
     //           |
     //             -> 6 -> 7
-    c.bench_function("side_branch", |b| {
-        b.iter_with_setup(
-            || {
-                let (chain, shared, dir) = new_chain();
-                let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
-                (0..5).for_each(|_| {
-                    let parent_index = blocks.len() - 1;
-                    gen_block(&mut blocks, parent_index);
-                });
-                (0..2).for_each(|i| {
-                    let parent_index = i + 2;
-                    gen_block(&mut blocks, parent_index);
-                });
-                blocks
-                    .clone()
-                    .into_iter()
-                    .skip(1)
-                    .take(5)
-                    .for_each(|block| {
+    c.bench_function_over_inputs(
+        "side_branch",
+        |b, txs_size| {
+            b.iter_with_setup(
+                || {
+                    let (chain, shared, dir, system_cell_hash, data_hash) = new_chain(*txs_size);
+                    let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
+                    (0..5).for_each(|_| {
+                        let parent_index = blocks.len() - 1;
+                        gen_block(&mut blocks, parent_index, &system_cell_hash, &data_hash);
+                    });
+                    (0..2).for_each(|i| {
+                        let parent_index = i + 2;
+                        gen_block(&mut blocks, parent_index, &system_cell_hash, &data_hash);
+                    });
+                    blocks
+                        .clone()
+                        .into_iter()
+                        .skip(1)
+                        .take(5)
+                        .for_each(|block| {
+                            chain
+                                .process_block(Arc::new(block))
+                                .expect("process block OK")
+                        });
+                    (chain, blocks, dir)
+                },
+                |(chain, blocks, _dir)| {
+                    blocks.into_iter().skip(6).for_each(|block| {
                         chain
                             .process_block(Arc::new(block))
                             .expect("process block OK")
                     });
-                (chain, blocks, dir)
-            },
-            |(chain, blocks, _dir)| {
-                blocks.into_iter().skip(6).for_each(|block| {
-                    chain
-                        .process_block(Arc::new(block))
-                        .expect("process block OK")
-                });
-            },
-        )
-    });
+                },
+            )
+        },
+        txs_sizes.clone(),
+    );
 
     // benchmark processing 4 blocks for switching fork
     // 0 -> 1 -> 2 -> 3 -> 4 -> 5
     //           |
     //             -> 6 -> 7 -> 8 -> 9
-    c.bench_function("switch_fork", |b| {
-        b.iter_with_setup(
-            || {
-                let (chain, shared, dir) = new_chain();
-                let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
-                (0..5).for_each(|_| {
-                    let parent_index = blocks.len() - 1;
-                    gen_block(&mut blocks, parent_index);
-                });
-                (0..4).for_each(|i| {
-                    let parent_index = i + 2;
-                    gen_block(&mut blocks, parent_index);
-                });
-                blocks
-                    .clone()
-                    .into_iter()
-                    .skip(1)
-                    .take(7)
-                    .for_each(|block| {
+    c.bench_function_over_inputs(
+        "switch_fork",
+        |b, txs_size| {
+            b.iter_with_setup(
+                || {
+                    let (chain, shared, dir, system_cell_hash, data_hash) = new_chain(*txs_size);
+                    let mut blocks = vec![shared.block(&shared.genesis_hash()).unwrap()];
+                    (0..5).for_each(|_| {
+                        let parent_index = blocks.len() - 1;
+                        gen_block(&mut blocks, parent_index, &system_cell_hash, &data_hash);
+                    });
+                    (0..4).for_each(|i| {
+                        let parent_index = i + 2;
+                        gen_block(&mut blocks, parent_index, &system_cell_hash, &data_hash);
+                    });
+                    blocks
+                        .clone()
+                        .into_iter()
+                        .skip(1)
+                        .take(7)
+                        .for_each(|block| {
+                            chain
+                                .process_block(Arc::new(block))
+                                .expect("process block OK")
+                        });
+                    (chain, blocks, dir)
+                },
+                |(chain, blocks, _dir)| {
+                    blocks.into_iter().skip(8).for_each(|block| {
                         chain
                             .process_block(Arc::new(block))
                             .expect("process block OK")
                     });
-                (chain, blocks, dir)
-            },
-            |(chain, blocks, _dir)| {
-                blocks.into_iter().skip(8).for_each(|block| {
-                    chain
-                        .process_block(Arc::new(block))
-                        .expect("process block OK")
-                });
-            },
-        )
-    });
+                },
+            )
+        },
+        txs_sizes.clone(),
+    );
 }
 
 criterion_group!(
     name = benches;
-    config = Criterion::default().sample_size(20);
+    config = Criterion::default().sample_size(10);
     targets = bench
 );
 criterion_main!(benches);
 
-fn new_chain() -> (
+fn new_chain(
+    txs_size: usize,
+) -> (
     ChainController,
     Shared<ChainKVStore<CacheDB<RocksDB>>>,
     TempDir,
+    H256,
+    H256,
 ) {
+    let always_success = include_bytes!("../../resource/specs/cells/always_success");
+    let cell_output = CellOutput::new(
+        Capacity::bytes(always_success.len()).unwrap(),
+        always_success.to_vec(),
+        Script::default(),
+        None,
+    );
+    let data_hash = cell_output.data_hash();
+
     let cellbase = TransactionBuilder::default()
         .input(CellInput::new_cellbase_input(0))
-        .output(CellOutput::new(
-            Capacity::zero(),
-            vec![],
-            Script::default(),
-            None,
-        ))
+        .output(cell_output)
         .build();
 
-    // create genesis block with 100 tx
-    let transactions: Vec<Transaction> = (0..100)
+    let system_cell_hash = cellbase.hash();
+
+    // create genesis block with N txs
+    let transactions: Vec<Transaction> = (0..txs_size)
         .map(|i| {
             TransactionBuilder::default()
                 .input(CellInput::new(OutPoint::null(), 0, vec![]))
                 .output(CellOutput::new(
                     capacity_bytes!(50_000),
-                    vec![i],
-                    Script::always_success(),
+                    i.to_le_bytes().to_vec(),
+                    Script::new(Vec::new(), data_hash.clone()),
                     None,
                 ))
                 .build()
@@ -164,7 +187,9 @@ fn new_chain() -> (
         .transactions(transactions)
         .with_header_builder(HeaderBuilder::default().difficulty(U256::from(1000u64)));
 
-    let consensus = Consensus::default().set_genesis_block(genesis_block);
+    let mut consensus = Consensus::default().set_genesis_block(genesis_block);
+    consensus.tx_proposal_window = ProposalWindow(1, 10);
+    consensus.cellbase_maturity = 0;
 
     let db_dir = tempdir().unwrap();
     let shared = SharedBuilder::<CacheDB<RocksDB>>::default()
@@ -176,10 +201,21 @@ fn new_chain() -> (
         .build();
     let notify = NotifyService::default().start::<&str>(None);
     let chain_service = ChainBuilder::new(shared.clone(), notify).build();
-    (chain_service.start::<&str>(None), shared, db_dir)
+    (
+        chain_service.start::<&str>(None),
+        shared,
+        db_dir,
+        system_cell_hash,
+        data_hash,
+    )
 }
 
-fn gen_block(blocks: &mut Vec<Block>, parent_index: usize) {
+fn gen_block(
+    blocks: &mut Vec<Block>,
+    parent_index: usize,
+    system_cell_hash: &H256,
+    data_hash: &H256,
+) {
     let p_block = &blocks[parent_index];
 
     let (number, timestamp, difficulty) = (
@@ -199,13 +235,13 @@ fn gen_block(blocks: &mut Vec<Block>, parent_index: usize) {
         .build();
 
     // spent n-2 block's tx and proposal n-1 block's tx
-    let transactions: Vec<Transaction> = if blocks.len() > parent_index + 1 {
+    let transactions: Vec<Transaction> = if blocks.len() > 1 {
         let pp_block = &blocks[parent_index - 1];
         pp_block
             .transactions()
             .iter()
             .skip(1)
-            .map(|tx| create_transaction(tx.hash()))
+            .map(|tx| create_transaction(tx.hash(), system_cell_hash, data_hash))
             .collect()
     } else {
         vec![]
@@ -215,7 +251,7 @@ fn gen_block(blocks: &mut Vec<Block>, parent_index: usize) {
         .transactions()
         .iter()
         .skip(1)
-        .map(|tx| create_transaction(tx.hash()).proposal_short_id())
+        .map(|tx| create_transaction(tx.hash(), system_cell_hash, data_hash).proposal_short_id())
         .collect();
 
     let block = BlockBuilder::default()
@@ -234,14 +270,15 @@ fn gen_block(blocks: &mut Vec<Block>, parent_index: usize) {
     blocks.push(block);
 }
 
-fn create_transaction(hash: H256) -> Transaction {
+fn create_transaction(parent_hash: H256, system_cell_hash: &H256, data_hash: &H256) -> Transaction {
     TransactionBuilder::default()
         .output(CellOutput::new(
             capacity_bytes!(50_000),
-            vec![],
-            Script::always_success(),
+            (0..255).collect(),
+            Script::new(vec![(0..255).collect()], data_hash.clone()),
             None,
         ))
-        .input(CellInput::new(OutPoint::new(hash, 0), 0, vec![]))
+        .input(CellInput::new(OutPoint::new(parent_hash, 0), 0, vec![]))
+        .dep(OutPoint::new(system_cell_hash.clone(), 0))
         .build()
 }