This PR refactor `BlockVerifier` to improve the performance of block processing by avoiding loading block body.

before:
```
main_branch/100         time:   [328.29 ms 360.15 ms 385.02 ms]                          
main_branch/200         time:   [728.01 ms 753.81 ms 774.04 ms]                          
main_branch/500         time:   [1.8385 s 1.8593 s 1.8964 s]                             
main_branch/1000        time:   [4.0649 s 4.1938 s 4.3535 s]     
```

after:
```
main_branch/100         time:   [301.37 ms 311.24 ms 324.40 ms]                          
                        change: [-14.707% -9.8664% -5.8876%] (p = 0.00 < 0.05)
main_branch/200         time:   [613.18 ms 616.63 ms 623.17 ms]                          
                        change: [-17.907% -15.483% -12.942%] (p = 0.00 < 0.05)
main_branch/500         time:   [1.6118 s 1.6272 s 1.6501 s]                             
                        change: [-14.673% -12.066% -9.2406%] (p = 0.00 < 0.05)
main_branch/1000        time:   [3.5405 s 3.6042 s 3.7018 s]                              
                        change: [-15.749% -12.350% -9.1919%] (p = 0.00 < 0.05)
```